To handle delete of multiplicity composite widget
git-svn-id: https://svn.kuali.org/repos/student/ks-core/branches/ks-core-dev@7738 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ ks-core-ui/src/main/java/org/kuali/student/core/organization/assembly/OrgOrgRelationAssembler.java
@@ -142,7 +142,9 @@ public class OrgOrgRelationAssembler implements Assembler<Data, OrgorgRelationHe
 //                OrgOrgRelationInfo orgOrgRelationInfo = buildOrgOrgRelationInfo(orgOrgRelation);
 //                orgOrgRelationInfo.setId(orgOrgRelation.getId());
                 try{
-                    StatusInfo  result = orgService.removeOrgOrgRelation(orgOrgRelation.getId());
+                    if(orgOrgRelation.getId()!=null){
+                        StatusInfo  result = orgService.removeOrgOrgRelation(orgOrgRelation.getId());
+                    }
                 }
                 catch(Exception e ){
                     e.printStackTrace();
